complete base conn interface
